requirements.txt
typer
lupa
pyyaml
requests
beautifulsoup4
